+++ example/CHANGELOG.md
@@ -1,5 +1,13 @@
 # example
 
+## 0.3.2
+
+### Patch Changes
+
+- fix sortable margin problem
+- Updated dependencies [undefined]
+  - @hexx/editor@0.4.2
+
 ## 0.3.1
 
 ### Patch Changes